+++ webpack.config.js
@@ -1,9 +1,19 @@
 const path = require("path");
 const Dotenv = require("dotenv-webpack");
 const HtmlWebpackPlugin = require("html-webpack-plugin");
+const WorkerPlugin = require("worker-plugin");
+const CopyWebpackPlugin = require("copy-webpack-plugin");
+
+let publicUrl = process.env.PUBLIC_URL || "/";
+if (!publicUrl.endsWith("/")) {
+  publicUrl += "/";
+}
 
 module.exports = {
-  entry: path.resolve(__dirname, "./src/web-component.js"),
+  entry: {
+    "web-component": path.resolve(__dirname, "./src/web-component.js"),
+    PyodideWorker: path.resolve(__dirname, "./src/PyodideWorker.js"),
+  },
   module: {
     rules: [
       {
@@ -28,9 +38,7 @@ module.exports = {
           {
             loader: "sass-loader",
             options: {
-              sassOptions: {
-                sourceMap: true,
-              },
+              sourceMap: true,
             },
           },
         ],
@@ -62,11 +70,17 @@ module.exports = {
     ],
   },
   resolve: {
-    extensions: ["*", ".js", ".jsx", ".css"],
+    extensions: [".*", ".js", ".jsx", ".css"],
+    fallback: {
+      path: require.resolve("path-browserify"),
+      url: require.resolve("url/"),
+    },
   },
   output: {
     path: path.resolve(__dirname, "./build"),
-    filename: "web-component.js",
+    filename: "[name].js",
+    publicPath: publicUrl,
+    workerPublicPath: publicUrl,
   },
   devServer: {
     host: "0.0.0.0",
@@ -82,17 +96,40 @@ module.exports = {
       "Access-Control-Allow-Methods": "GET, POST, PUT, DELETE, PATCH, OPTIONS",
       "Access-Control-Allow-Headers":
         "X-Requested-With, content-type, Authorization",
+      // Pyodide - required for input and code interruption - needed on the host app
+      "Cross-Origin-Opener-Policy": "same-origin",
+      "Cross-Origin-Embedder-Policy": "require-corp",
+    },
+    setupMiddlewares: (middlewares, devServer) => {
+      devServer.app.use((req, res, next) => {
+        // PyodideWorker scripts - cross origin required on scripts needed for importScripts
+        if (
+          [
+            "/pyodide/shims/_internal_sense_hat.js",
+            "/pyodide/shims/pygal.js",
+            "/PyodideWorker.js",
+          ].includes(req.url)
+        ) {
+          res.setHeader("Cross-Origin-Resource-Policy", "cross-origin");
+        }
+        next();
+      });
+      return middlewares;
     },
   },
   plugins: [
+    new WorkerPlugin(),
     new Dotenv({
-      path: "./.env.webcomponent",
+      path: "./.env",
       systemvars: true,
     }),
     new HtmlWebpackPlugin({
       inject: "body",
       template: "src/web-component.html",
       filename: "web-component.html",
+    }),
+    new CopyWebpackPlugin({
+      patterns: [{ from: "public", to: "" }],
     }),
   ],
   stats: "minimal",